make running einops tutorial locally easier
